app/requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
autopep8==2.3.2
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.7
elastic-transport==9.1.0
elasticsearch==9.1.0
filelock==3.19.1
flake8==7.3.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jwt==1.4.0
MarkupSafe==3.0.2
mccabe==0.7.0
mpmath==1.3.0
mypy==1.17.1
mypy_extensions==1.1.0
narwhals==2.4.0
networkx==3.5
numpy==2.3.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
pillow==11.3.0
protobuf==6.32.0
pyarrow==21.0.0
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pyflakes==3.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
six==1.17.0
smmap==5.0.2
streamlit==1.49.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.0
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.1
triton==3.4.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
